optimisation-worker/requirements.txt
rq
redis
python-dotenv
sqlalchemy
psycopg2